getcomponentDescriptionDTOs safety
