+++ compendium/DeclarativeServices/src/ServiceComponentRuntimeImpl.cpp
@@ -133,7 +133,20 @@ namespace cppmicroservices::scrimpl
         std::vector<ComponentDescriptionDTO> componentDTOs;
         for (auto holder : compMgrs)
         {
-            componentDTOs.push_back(CreateDTO(holder));
+            try
+            {
+                componentDTOs.push_back(CreateDTO(holder));
+            }
+            catch (std::exception const& e)
+            {
+                logger->Log(cppmicroservices::logservice::SeverityLevel::LOG_DEBUG,
+                           "Failed to create ComponentDescriptionDTO: " + std::string(e.what()));
+            }
+            catch (...)
+            {
+                logger->Log(cppmicroservices::logservice::SeverityLevel::LOG_DEBUG,
+                           "Failed to create ComponentDescriptionDTO: Unknown exception");
+            }
         }
         return componentDTOs;
     }
