+++ compendium/DeclarativeServices/test/gtest/TestAsyncWorkService.cpp
@@ -422,185 +422,4 @@ namespace test
 
         fut.get();
     }
-
-    /**
-     * Verify that a service's configuration can be updated from with a 'Modified' callback without
-     * deadlocking the framework trying to get the lock in configNotifier
-     */
-    TEST_F(TestAsyncWorkServiceEndToEnd, testUpdateConfigFromWithinModifiedCallback)
-    {
-        auto const& param = std::make_shared<AsyncWorkServiceThreadPool>(10);
-        auto context = framework.GetBundleContext();
-
-        // ASYNCWORKSERVICE
-        auto reg = context.RegisterService<cppmicroservices::async::AsyncWorkService>(param);
-
-        // CA
-        ::test::InstallAndStartConfigAdmin(context);
-
-        //  CA Service
-        auto CARef = context.GetServiceReference<cppmicroservices::service::cm::ConfigurationAdmin>();
-        auto configAdminService = context.GetService<cppmicroservices::service::cm::ConfigurationAdmin>(CARef);
-        ASSERT_TRUE(configAdminService) << "GetService failed for ConfigurationAdmin";
-
-        // Start bundle
-        std::string componentName = "sample::ServiceComponentCA10";
-
-        auto testBundle = ::test::InstallAndStartBundle(context, "TestBundleDSCA10");
-        ::test::InstallAndStartBundle(context, "TestBundleDSCA03");
-
-        auto configObject = configAdminService->GetConfiguration(componentName);
-
-        auto props = configObject->GetProperties();
-        auto configObject2 = configAdminService->GetConfiguration("sample::ServiceComponentCA02");
-
-        props["uniqueProp"] = std::string("UNIQUE");
-        auto fut = configObject->Update(props);
-        fut.get();
-        fut = configObject2->Update(props);
-        fut.get();
-        props["configID"] = std::string("sample::ServiceComponentCA03");
-        fut = configObject->Update(props);
-        fut.get();
-
-        // GetService to make component active
-        auto serviceRef = context.GetServiceReference<test::CAInterface>();
-        auto service = context.GetService<test::CAInterface>(serviceRef);
-        ASSERT_TRUE(service) << "GetService failed for CAInterface";
-
-        ASSERT_NO_THROW(reg.Unregister());
-    }
-
-    TEST_F(TestAsyncWorkServiceEndToEnd, TestConcurrentBundleInstalls)
-    {
-        auto const& param = std::make_shared<AsyncWorkServiceThreadPool>(20);
-        auto ctx = framework.GetBundleContext();
-        auto reg = ctx.RegisterService<cppmicroservices::async::AsyncWorkService>(param);
-
-        auto paths = ::test::GetAllTestBundleLocations();
-
-        constexpr int numThreads = 18;
-        Barrier b(numThreads);
-        std::vector<std::future<void>> futures;
-        std::atomic<bool> failed = false;
-        for (int i = 0; i < numThreads; ++i)
-        {
-            futures.emplace_back(
-                std::async(std::launch::async,
-                           [&]()
-                           {
-                               b.Wait();
-                               auto threadCtx = framework.GetBundleContext();
-
-                               std::map<std::string, std::string> installedBundles;
-
-                               for (auto const& path : paths)
-                               {
-                                   std::vector<cppmicroservices::Bundle> bundles;
-                                   try
-                                   {
-                                       bundles = threadCtx.InstallBundles(path); // returns std::vector<Bundle>
-                                   }
-                                   catch (...)
-                                   {
-                                       // ignore malformed bundles
-                                       continue;
-                                   }
-                                   for (auto const& bundle : bundles)
-                                   {
-                                       std::string name = bundle.GetSymbolicName();
-                                       auto const [preexistingBundle, newBundleWasInserted]
-                                           = installedBundles.try_emplace(name, path);
-                                       if (!newBundleWasInserted)
-                                       {
-                                           std::cout << "Bundle '" << name << "' already installed from '"
-                                                     << preexistingBundle->second << "'. New install attempt from '"
-                                                     << path << "'." << std::endl;
-                                           failed.store(true);
-                                       }
-                                   }
-                               }
-                           }));
-        }
-
-        // Wait for all threads to complete
-        for (auto& fut : futures)
-        {
-            fut.get();
-        }
-
-        ASSERT_FALSE(failed.load());
-
-        framework.Stop();
-        framework.WaitForStop(std::chrono::milliseconds::zero());
-    }
-
-    class basicDS5 : public test::DSGraph05
-    {
-      public:
-        basicDS5() = default;
-        ~basicDS5() override = default;
-        std::string
-        Description() override
-        {
-            return "basicDS5";
-        };
-    };
-
-    TEST_F(TestAsyncWorkServiceEndToEnd, VerifyNoDuplicateDeactivate)
-    {
-        auto ctx = framework.GetBundleContext();
-        auto reg = ctx.RegisterService<cppmicroservices::async::AsyncWorkService>(
-            std::make_shared<AsyncWorkServiceThreadPool>(20));
-
-        // install And start referencING service
-        auto path = "DSGraph04";
-        auto bundle = ::test::InstallAndStartBundle(ctx, path);
-
-        // create referencED service instance
-        auto initialRefdSvcReg = ctx.RegisterService<test::DSGraph05>(std::make_shared<test::basicDS5>());
-
-        auto sRef = ctx.GetServiceReference<test::DSGraph04>();
-        ASSERT_TRUE(sRef);
-        auto svc = ctx.GetService<test::DSGraph04>(sRef);
-
-        ASSERT_EQ(svc->Description(), "DSGraph04");
-
-        constexpr int numThreads = 2;
-        Barrier beforeStart(numThreads);
-        std::vector<std::future<void>> futures;
-
-        futures.emplace_back(std::async(std::launch::async,
-                                        [&]()
-                                        {
-                                            beforeStart.Wait();
-                                            initialRefdSvcReg.Unregister();
-                                        }));
-
-        futures.emplace_back(
-            std::async(std::launch::async,
-                       [&]()
-                       {
-                           beforeStart.Wait();
-
-                           auto props = cppmicroservices::ServiceProperties({
-                               { std::string("component.name"), std::string("test1") }
-                           });
-
-                           auto reg = ctx.RegisterService<test::DSGraph05>(std::make_shared<test::basicDS5>(), props);
-                       }));
-
-        for (auto& fut : futures)
-        {
-            fut.get();
-        }
-
-        sRef = ctx.GetServiceReference<test::DSGraph04>();
-        ASSERT_TRUE(sRef);
-        svc = ctx.GetService<test::DSGraph04>(sRef);
-        ASSERT_EQ(svc->Description(), "DSGraph04");
-
-        framework.Stop();
-        framework.WaitForStop(std::chrono::milliseconds::zero());
-    }
 }; // namespace test
